login.component.spec.ts
This test creates an instance of the LoginComponent, sets the email property, calls the onSubmit() function, and checks if the email property is still set to the same value after the function is called. Note that this test does not require any Angular testing utilities since we are not testing the component's template or any Angular-specific functionality.